server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
chromadb>=0.5.0
sentence-transformers>=3.2.0
torch==2.2.2+cu121
torchvision==0.17.2+cu121
torchaudio==2.2.2+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
pypdf>=5.0.0
python-docx>=1.1.0
httpx>=0.27.0
orjson>=3.10.0
numpy==1.26.4
pypdf
